requirements.txt
# Core Web Framework
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0

# Database
psycopg2-binary>=2.9.0

# Scheduling and Background Tasks  
apscheduler>=3.10.0

# Configuration and Environment
python-dotenv>=1.0.0

# HTTP Requests and API Integration
requests>=2.31.0

# Data Processing and Analysis
numpy>=1.24.0
scipy>=1.11.0

# Machine Learning and Predictive Modeling
scikit-learn>=1.3.0
statsmodels>=0.14.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0